requirements.txt
python-dotenv
requests
streamlit
pandas
openai
python-telegram-bot